# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/public/dapp/banner/list": {
            "post": {
                "description": "获取轮播图列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页"
                ],
                "summary": "获取轮播图列表",
                "operationId": "GetBannerList",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.GetBannerListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.GetBannerListResp"
                        }
                    }
                }
            }
        },
...
        "/api/v1/public/dapp/network/list": {
            "post": {
                "description": "获取网络列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页"
                ],
                "summary": "获取网络列表",
                "operationId": "GetNetworkList",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.GetNetworkListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.GetNetworkListResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/tool/categories/list": {
            "post": {
                "description": "获取工具分类列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页"
                ],
                "summary": "获取工具分类列表",
                "operationId": "GetDiscoverToolCategoriesList",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.GetDiscoverToolCategoriesListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.GetDiscoverToolCategoriesListResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/tool/event/list": {
            "post": {
                "description": "发现页工具访问记录列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页"
                ],
                "summary": "发现页工具访问记录列表",
                "operationId": "GetDiscoverToolEventList",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.GetDiscoverToolEventListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.GetDiscoverToolEventListResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/tool/event/operation": {
            "post": {
                "description": "发现页工具访问记录操作",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页"
                ],
                "summary": "发现页工具访问记录操作",
                "operationId": "DiscoverToolEventOperation",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.DiscoverToolEventOperationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.DiscoverToolEventOperationResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/tool/favorites/list": {
            "post": {
                "description": "获取发现页工具收藏列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页"
                ],
                "summary": "获取发现页工具收藏列表",
                "operationId": "GetDiscoverToolFavoritesList",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.GetDiscoverToolFavoritesListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.GetDiscoverToolFavoritesListResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/tool/favorites/operation": {
            "post": {
                "description": "发现页工具收藏操作",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页"
                ],
                "summary": "发现页工具收藏操作",
                "operationId": "DiscoverToolFavoritesOperation",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.DiscoverToolFavoritesOperationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.DiscoverToolFavoritesOperationResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/tool/info/list": {
            "post": {
                "description": "获取工具列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页"
                ],
                "summary": "获取工具列表",
                "operationId": "GetDiscoverToolInfoList",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.GetDiscoverToolInfoListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.GetDiscoverToolInfoListResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/toolbar/list": {
            "post": {
                "description": "获取发现页工具栏列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页"
                ],
                "summary": "获取发现页工具栏列表",
                "operationId": "GetDiscoverToolbarList",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.GetDiscoverToolbarListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.GetDiscoverToolbarListResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "protos.DiscoverToolEventOperationReq": {
            "type": "object",
            "properties": {
                "event_type": {
                    "description": "view/click/share/fav/comment",
                    "type": "string"
                },
                "tool_id": {
                    "type": "integer"
                }
            }
        },
...
        "protos.GetBannerListReq": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "protos.GetBannerListResp": {
            "type": "object",
            "properties": {
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "expiration_time": {
                    "description": "过期时间",
                    "type": "string"
                },
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                },
                "img_href": {
                    "description": "图片地址",
                    "type": "string"
                },
                "need_login": {
                    "description": "是否需要登录：1-不需要，2-需要",
                    "type": "integer"
                },
                "position": {
                    "description": "位置 key",
                    "type": "string"
                },
                "skip_target": {
                    "description": "跳转目标：1-H5,2-客服",
                    "type": "integer"
                },
                "skip_url": {
                    "description": "跳转URL",
                    "type": "string"
                },
                "sort": {
                    "description": "顺序",
                    "type": "integer"
                },
                "start_at": {
                    "description": "开始时间",
                    "type": "string"
                },
                "status": {
                    "description": "状态：1-上架，2-下架",
                    "type": "integer"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "updated_time": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "protos.GetDiscoverToolCategoriesListReq": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "protos.GetDiscoverToolCategoriesListResp": {
            "type": "object",
            "properties": {
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "description": {
                    "description": "说明",
                    "type": "string"
                },
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                },
                "img_href": {
                    "description": "图片地址",
                    "type": "string"
                },
                "is_show": {
                    "description": "1-显示，2-隐藏",
                    "type": "integer"
                },
                "skip_url": {
                    "description": "跳转URL",
                    "type": "string"
                },
                "sort": {
                    "description": "顺序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1-上架，2-下架",
                    "type": "integer"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "updated_time": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "protos.GetNetworkListReq": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "protos.GetNetworkListResp": {
            "type": "object",
            "properties": {
                "chain_id": {
                    "description": "链ID，如 1、56、solana",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "explorer_url": {
                    "description": "浏览器地址，例如 https://bscscan.com",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "description": "图标",
                    "type": "string"
                },
                "name": {
                    "description": "网络名称，如 Ethereum、BSC、Solana",
                    "type": "string"
                },
                "rpc_url": {
                    "description": "默认 RPC，非必填",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态 1启用 2禁用",
                    "type": "integer"
                },
                "symbol": {
                    "description": "链主币符号，例如 ETH、BNB、SOL",
                    "type": "string"
                }
            }
        },
...
    }
}